requirements.txt
# Deep Learning
torch>=2.2.0
torchvision>=0.17.0
pytorch-lightning>=2.2.0
tensorboard>=2.15.0

# Data Processing & Math
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Visualization & Logging
matplotlib>=3.8.0
seaborn>=0.13.0
rich>=13.7.0
tqdm>=4.66.0
ipywidgets>=8.1.0  # For tqdm in Notebooks

# Utilities
pyyaml>=6.0.1
pillow>=10.1.0
packaging>=23.2

# Testing
pytest>=8.0.0